+++ src/app/ngx-editor/common/services/command-executor.ts
