+++ src/app/ngx-editor/ngx-editor.module.ts
@@ -5,7 +5,7 @@ import { NgxGrippieComponent } from './ngx-grippie/ngx-grippie.component';
 import { NgxToolbarComponent } from './ngx-toolbar/ngx-toolbar.component';
 import { NgxEditorMessageComponent } from './ngx-editor-message/ngx-editor-message.component';
 import { MessageService } from './ngx-editor-message/message.service';
-import { CommandExecutor } from './command-executor';
+import { CommandExecutor } from './common/services/command-executor';
 
 @NgModule({
   imports: [
