+++ src/app/ngx-editor/ngx-grippie/ngx-grippie.component.spec.ts
@@ -4,7 +4,7 @@ import { ElementRef } from '@angular/core';
 import { NgxGrippieComponent } from './ngx-grippie.component';
 import { NgxEditorComponent } from '../ngx-editor.component';
 import { MessageService } from '../ngx-editor-message/message.service';
-import { CommandExecutor } from '../command-executor';
+import { CommandExecutor } from '../common/services/command-executor';
 
 export class MockElementRef extends ElementRef {
   constructor() { super(null); }
